+++ test/pine/hints_test.clj
@@ -71,13 +71,15 @@
              :pine "of: y.employee .reports_to"}]
            (-> "employee | of: employee" gen :table))))
 
-  (testing "Generate `select` hints with relevant columns"
+  (testing "Generate `select` hints with columns specified"
     (is (= []
            (-> "x.company | s: does_not_exist" gen :select)))
     (is (= [{:column "id" :alias "c_0"}]
            (-> "x.company | s: i" gen :select)))
     (is (= ["company_id" "id"] ;;  "reports_to" is not returned
-           (->> "y.employee | s: id" gen :select (map :column)))))
+           (->> "y.employee | s: id" gen :select (map :column))))
+    (is (= ["reports_to" "company_id" "id"]
+           (->> "y.employee as e | s: e.*" gen :select (map :column)))))
 
   (testing "Generate `select-partial` hints"
     (is (= [{:column "id" :alias "c_0"}]     (->  "company    | s:"                 gen :select)))